Normalized data in customer data
PCA calculates a new projection for the data based on the standard deviation of the variables. If the data is not normalized, the feature with a high standard deviation will be weighted heavily in the calculation of a new axis. Normalization ensures that all variables have the same standard deviation which will allow for the calculation of a more accurate projection